separate tokenizer functions from main.c
token making functions and tokenizer is in tokenize.c
