+++ code/codegenerator.c
@@ -1,6 +1,6 @@
 #include "cc.h"
-#include<stdio.h>
-#include<stdlib.h>
+//#include<stdio.h>
+//#include<stdlib.h>
